Minor refactoring. Fix signal handling
