+++ cmd/handlers/actions.go
@@ -1,0 +1,9 @@
+// Copyright (C) 2025  Anton Kachurin
+package handlers
+
+// ActionHandlers handle CLI actions.
+type ActionHandlers struct {
+	filePath, repoPath, configPath string
+}
+
+const DefaultCommandName = commandNameList
